fix: saslAuthenticate
TODO: use opaque packets if SaslHandshakeRequest version is v0

If SaslHandshakeRequest version is v0, a series of SASL client and server tokens corresponding to the mechanism are sent as opaque packets without wrapping the messages with Kafka protocol headers. If SaslHandshakeRequest version is v1, the SaslAuthenticate request/response are used, where the actual SASL tokens are wrapped in the Kafka protocol.

refer to https://kafka.apache.org/protocol#sasl_handshake